+++ src/main/java/org/example/travelplanner/controller/PlanController.java
@@ -1,6 +1,6 @@
 package org.example.travelplanner.controller;
 
-import org.example.travelplanner.entity.Plan;
+import org.example.travelplanner.Plan;
 import org.example.travelplanner.service.PlanService;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
@@ -21,7 +21,7 @@ public class PlanController {
     }
 
     @GetMapping("/{id}")
-    public Plan getById(@PathVariable Long id) {
+    public Plan getById(@PathVariable int id) {
         return planService.getPlanById(id);
     }
 
@@ -31,12 +31,12 @@ public class PlanController {
     }
 
     @PutMapping("/{id}")
-    public Plan update(@PathVariable Long id, @RequestBody Plan plan) {
+    public Plan update(@PathVariable int id, @RequestBody Plan plan) {
         return planService.updatePlan(id, plan);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable int id) {
         planService.deletePlan(id);
     }
 }
