+++ src/main/java/org/example/travelplanner/repository/AttractionRepository.java
@@ -3,5 +3,5 @@ package org.example.travelplanner.repository;
 import org.example.travelplanner.Attraction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttractionRepository extends JpaRepository<Attraction, Long> {
+public interface AttractionRepository extends JpaRepository<Attraction, Integer> {
 }
