+++ src/main/java/org/example/travelplanner/repository/DestinationRepository.java
@@ -3,5 +3,5 @@ package org.example.travelplanner.repository;
 import org.example.travelplanner.Destination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DestinationRepository extends JpaRepository<Destination, Long> {
+public interface DestinationRepository extends JpaRepository<Destination, Integer> {
 }
