+++ src/main/java/org/example/travelplanner/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package org.example.travelplanner.repository;
 import org.example.travelplanner.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
 }
